+++ CHANGELOG.md
@@ -5,6 +5,13 @@ Headlines: Added, Changed, Deprecated, Removed, Fixed, Security
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Removed
+
+- `options.noErrors` is removed. Use `options.errors` instead.
+- The virtually unused `meta` has been removed. Use the Zod schema directly instead for reflection.
+
 ## [0.8.5] - 2023-05-04
 
 ### Fixed